[ALPS03521190] PMIC: fix MT6370 pd30 build error in 32-bit project
[Detail]
fix MT6370 pd30 build error in 32-bit project

MTK-Commit-Id: 4e5420c2045cfdde7d2c7a642d9f783df99bde97

CR-Id: ALPS03521190
Feature: Power Management

+++ drivers/misc/mediatek/usb_c/tcpc/tcpci.c
@@ -18,7 +18,7 @@
 
 static int tcpc_check_notify_time(struct timeval *begin, struct timeval *end)
 {
-	return (timeval_to_ns(end) - timeval_to_ns(begin))/1000/1000;
+	return (int)(timeval_to_ns(end) - timeval_to_ns(begin))/1000/1000;
 }
 
 int tcpci_check_vbus_valid_from_ic(struct tcpc_device *tcpc)
